WIP: Condense examples
The current examples are partitioned into multiple examples block. E.g. typing

'import numpy'
'one'

into the terminal yields 4 examples, and only two of the example blocks in 'numpy.ones_like'
actually have that function being called. It's clear that this is really two examples
in four blocks, so it'd be nice if we squashed blocks together until we get the right function call.